Revert changes from master up to v1.1 (37837a25e089a5b6356554b9861634dd4afd3abb)

+++ dfd/dfdAlgorithm/src/fdiscovery/columns/AgreeSet.java
@@ -10,7 +10,7 @@ public class AgreeSet extends ColumnCollection {
 
 	private static final long serialVersionUID = -5335032949377336772L;
 
-	public AgreeSet(Set<Point> set1, Set<Point> set2, int numberOfColumns) {
+	public AgreeSet(Set<Point> set1, Set<Point> set2, long numberOfColumns) {
 		super(numberOfColumns);
 		Set<Point> intersected = Sets.intersection(set1, set2);
 		for (Point columnToIdentifier : intersected) {
